Move existing services to new folder
We also rename the config for the process as it is no longer a job.
