+++ .env.example
@@ -61,8 +61,8 @@ REDIS_DISABLE_TLS=false
 BILLING_ANNUAL_BATCH_SIZE=5
 BILLING_WAIT_FOR_STATUS_PAUSE_IN_MS=1000
 
-# Jobs config
-JOBS_LICENCE_CHANGES_BATCH_SIZE=10
+# Licences config
+LICENCE_END_DATES_BATCH_SIZE=10
 
 # Cookie secret for authenticating requests coming via the UI. Note that the value should be the same as `COOKIE_SECRET`
 # on the UI side
