+++ app/services/licences/end-dates/determine-earliest-licence-changed-date.service.js
